pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "phased-array-systems"
dynamic = ["version"]
description = "Phased array antenna system design, optimization, and performance visualization"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "phased-array-systems contributors" },
]
keywords = [
    "phased array",
    "antenna",
    "radar",
    "communications",
    "RF",
    "system design",
    "trade study",
    "optimization",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "phased-array-modeling>=1.3.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pydantic>=2.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pandas-stubs>=2.0",
]
plotting = [
    "plotly>=5.0",
    "kaleido>=0.2",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.4",
    "mkdocstrings[python]>=0.24",
    "mkdocs-jupyter>=0.24",
    "mkdocs-gen-files>=0.5",
    "mkdocs-literate-nav>=0.6",
    "mkdocs-section-index>=0.3",
]

[project.scripts]
pasys = "phased_array_systems.cli:main"

[project.urls]
Homepage = "https://github.com/jman4162/phased-array-systems"
Documentation = "https://jman4162.github.io/phased-array-systems"
Repository = "https://github.com/jman4162/phased-array-systems"
Issues = "https://github.com/jman4162/phased-array-systems/issues"
Changelog = "https://jman4162.github.io/phased-array-systems/changelog"

[tool.hatch.version]
path = "src/phased_array_systems/__about__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/phased_array_systems"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
